fix sig/slog connection
svn path=/trunk/koffice/; revision=1062705

+++ krita/plugins/paintops/spray/kis_sprayop_option.cpp
@@ -43,7 +43,7 @@ KisSprayOpOption::KisSprayOpOption()
     connect(m_options->jitterMovementSpin, SIGNAL(valueChanged(double)), SIGNAL(sigSettingChanged()));
     connect(m_options->spacingSpin, SIGNAL(valueChanged(double)), SIGNAL(sigSettingChanged()));
     connect(m_options->scaleSpin, SIGNAL(valueChanged(double)), SIGNAL(sigSettingChanged()));
-    connect(m_options->particlesSpinBox, SIGNAL(valueChanged(int)), SIGNAL(sigSettingChanged()));
+    connect(m_options->particlesSpinBox, SIGNAL(valueChanged(double)), SIGNAL(sigSettingChanged()));
     connect(m_options->densityChBox, SIGNAL(stateChanged(int)), SIGNAL(sigSettingChanged()));
 
     setConfigurationPage(m_options);
